Moved keras_download.py to subfolder within scripts
